refactor: move login factories to a folder
